[#9] Add a list of available operators to the README
https://github.com/rxjsx/rxjsx/issues/9
Currently, descriptive documentation of the available operators exists in a deeper README file. We should create a table on the main README that contains all the operators in rows with links to their descriptions in the other README file.

The table can have only one column in the first PR. A second column with short descriptions can be added in a separate pull request.